Support jvm version libhdfs in velox (#9835)
Summary:
Currently, Gluten will throw hdfs connection failures when executing queries if their HDFS system employs Kerberos authentication and Viewfs support. This is due to the fact that the existing libhdfs3 API does not support Kerberos authentication, whereas the JVM version of libhdfs is capable of invoking APIs that support Kerberos authentication. If the user's system has the `HADOOP_HOME `environment variable set, the JVM version of libhdfs will be used during the compilation of Gluten; if not set, the default libhdfs3 will be used instead.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9835

Reviewed By: xiaoxmeng, DanielHunte

Differential Revision: D64872596

Pulled By: pedroerp

fbshipit-source-id: 995ee73a9f8474f8a6467b926a56246073fee75e